+++ src/generics/Main.java
@@ -1,0 +1,18 @@
+package generics;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public class Main {
+    public static void main(String[] args) {
+        List<String> strings = new ArrayList<>();
+        strings.add("aergaer");
+        strings.clear();
+        Map<String, Object> hashMap = new HashMap<>();
+        hashMap.put("abc", ");
+        hashMap.get();
+
+    }
+}
