+++ src/generics/PetOwner.java
@@ -1,0 +1,7 @@
+package generics;
+
+import inheritance.Animal;
+
+public class PetOwner<T extends Animal> {
+
+}
