+++ src/inheritance/Main.java
@@ -1,4 +1,4 @@
-package Inheritance;
+package inheritance;
 
 public class Main {
     public static void main(String[] args) {
