+++ src/inheritance/Sloth.java
@@ -1,4 +1,4 @@
-package Inheritance;
+package inheritance;
 
 public class Sloth extends Animal {
