+++ src/inheritance/Speaker.java
@@ -1,4 +1,4 @@
-package Inheritance;
+package inheritance;
 
 public interface Speaker {
     void speak();
